requirements.txt
streamlit==1.29.0
nltk==3.8.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
python-whois==0.8.0
trafilatura==1.6.1
plotly==5.18.0
numpy==1.26.2
scikit-learn==1.3.2
